completed drawing triangles

+++ linedrawing.h
@@ -2,9 +2,9 @@
 #include "bitmap.h"
 
 struct vertex {
-    float x;
-    float y;
-    float z;
+    int x;
+    int y;
+    int z;
 };
 
 unsigned int normalizeCoords(float n);
